Changed the type of the variables distance and speed from float to uint32_t in activityClassifier.h.

+++ src/boot/ksdk1.1.0/activityClassifier.h
@@ -45,12 +45,12 @@ uint32_t LPFBuffer[BUFFER_SIZE] = {0}; // Initialised to 0.
 uint32_t maximumValue, minimumValue; // Not set until the classification algorithm runs.
 uint32_t LPFBufferMidpoint; // Not set until the classification algorithm runs.
 uint16_t numberOfCrossings = 0; // The number of times the midpoint of the low-pass filtered signal is crossed.
-uint16_t numberOfInflectionPoints = 0; // The  number of local maxima and minima in the array. Use (float)(numberOfInflectionPoints / 2) instead of the number of steps to avoid rounding errors.
+uint16_t numberOfInflectionPoints = 0; // The number of local maxima and minima in the array. Use numberOfInflectionPoints / 2 instead of the number of steps to avoid rounding errors.
 uint16_t cumulativeNumberOfInflectionPoints = 0; // The cumulative number of inflection points since starting the program.
 uint16_t cumulativeNumberOfSteps = 0; // Cumulative number of steps since booting the device. Initialised to 0.
 
-float distance = 0; // Initialised to 0.
-float speed = 0; // Initialised to 0.
+uint32_t distance = 0; // Initialised to 0.
+uint32_t speed = 0; // Initialised to 0.
 
 // The LPF uses a finite impulse response (FIR) structure.
 // The FIR coefficients h(n) are defined below.
